Move request from BaseCRUD to RequestMethods
 * Base objects do not necessarily need the request method

Signed off by: Diego Araújo <diegoamc90@gmail.com>
